gnu_linux_box/requirements.txt
torch==1.9.0
Pillow==8.4.0
protobuf==3.19.1
termcolor
pyaudio
google-api-python-client
google-cloud-speech
tts
playsound
vext
fuzzysearch
webrtcvad
wikipedia==1.4.0
spacy==3.1.1
six==1.16.0
requests==2.26.0
google-cloud-language
google-cloud-translate
pycrypto
flask
flask-restful
flask-cors
duckduckgo_search
opengraph_py3
